+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- SPARQL export automatically prefixes variables with `?` where appropriate.
 - SPARQL output uses native `IN`/`NOT IN` syntax where appropriate.
 
 ## [v8.16.0] - 2026-05-04